# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nutrition Platform API
  version: 0.1.0
  description: Versioned REST contract for the Brazil-first nutrition practice platform.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/patient-invitations:
    post:
      tags: [Patient Invitations]
      operationId: createPatientInvitation
      summary: Create an Organization-scoped patient invitation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientInvitationRequest'
      responses:
        '201':
          description: Invitation created; the raw capability token is returned once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientInvitationCreated'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
  /api/v1/patient-invitations/{token}/accept:
    post:
      tags: [Patient Invitations]
      operationId: acceptPatientInvitation
      summary: Accept an invitation using the authenticated patient's verified email
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientInvitationAcceptanceRequest'
      responses:
        '200':
          description: Care relationship activated and consent recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientInvitationAcceptanceResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: Request failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CreatePatientInvitationRequest:
      type: object
      additionalProperties: false
      required: [displayName, email]
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 160
        email:
          type: string
          format: email
          maxLength: 254
        careFocus:
          type: [string, 'null']
          maxLength: 120
    PatientInvitationCreated:
      type: object
      required: [invitationId, patientId, token, status, expiresAt]
      properties:
        invitationId:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        token:
          type: string
        status:
          type: string
          enum: [PENDING]
        expiresAt:
          type: string
          format: date-time
...
    PatientInvitationAcceptanceRequest:
      type: object
      additionalProperties: false
      required: [consentTextVersion]
      properties:
        consentTextVersion:
          type: string
          pattern: '^[a-z0-9-]{3,60}$'
    PatientInvitationAcceptanceResponse:
      type: object
      required: [patientId, relationshipStatus, acceptedAt]
      properties:
        patientId:
          type: string
          format: uuid
        relationshipStatus:
          type: string
          enum: [ACTIVE]
        acceptedAt:
          type: string
          format: date-time
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri